+++ src/lib/webext.ts
@@ -72,6 +72,17 @@ export async function activeTabContainerId() {
     return (await activeTab()).cookieStoreId
 }
 
+//#background_helper
+export async function activeTabContainer() {
+    let containerId = await activeTabContainerId()
+    if (containerId !== "firefox-default")
+        return await browserBg.contextualIdentities.get(containerId)
+    else
+        throw new Error(
+            "firefox-default is not a valid contextualIdentity (activeTabContainer)",
+        )
+}
+
 /** Compare major firefox versions */
 export async function firefoxVersionAtLeast(desiredmajor: number) {
     const versionstr = (await browserBg.runtime.getBrowserInfo()).version